:hammer: Add no ID guard clause
This makes no difference to the code here, in terms of what it's testing,
but I always like to try and make code fully-passing when it comes to
keeping pyright happy. This removes one (valid) complaint.

Still got one left though. It doesn't affect how the code runs, but it's
there and can't be unseen.