:sparkles: Example of a one-or-many-line input
Someone on Discord was doing a `Input` or `TextArea` widget thing, where a
single line would be an `Input` and more than one line would then become a
`TextArea`. This feels like way too much work to me, and I wrote this as an
example of how you might be *way* more efficient while achieving the same
layout aim.