:sparkles: Add an example of messages up, attributes down
For a question on Discord. The person asking seemed confused by the
"coordinating widgets" section in the docs, seemingly not being able to see
how the advice given and how the example that was all about widgets could
relate to widgets being coordinated by a screen or an app. So I wrote this
very quickly to hopefully illustrate that aspect of it.

Apologies to anyone who knows that names don't always come in "first" and
"last" portions; I was just stuck for a simple "complex" example.