:sparkles: Add a test for multiple visible dialogs
It's not intended that anyone use it this way; it's questionable that it's
even allowed really; but the spec is that this is a widget and not a screen,
so if Dialog is a widget that then implies that there an be multiple ones on
the screen at once.

So... let's test that weird situation.